move layouts into layout directory
