+++ src/Components/Publishing/Layouts/__tests__/SeriesLayout.test.tsx
@@ -2,14 +2,14 @@ import "jest-styled-components"
 import React from "react"
 import renderer from "react-test-renderer"
 import { SeriesArticle, SeriesArticleSponsored } from "../../Fixtures/Articles"
-import { Series } from "../Series"
+import { SeriesLayout } from "../SeriesLayout"
 
 it("renders a series properly", () => {
-  const series = renderer.create(<Series article={SeriesArticle} />).toJSON()
+  const series = renderer.create(<SeriesLayout article={SeriesArticle} />).toJSON()
   expect(series).toMatchSnapshot()
 })
 
 it("renders a sponsored series properly", () => {
-  const series = renderer.create(<Series article={SeriesArticleSponsored} />).toJSON()
+  const series = renderer.create(<SeriesLayout article={SeriesArticleSponsored} />).toJSON()
   expect(series).toMatchSnapshot()
 })
