+++ src/Components/Publishing/Typings.ts
@@ -1,4 +1,4 @@
-export type Layout = "classic" | "standard" | "feature"
+export type Layout = "classic" | "feature" | "series" | "standard" | "video"
 export type SectionLayout = "blockquote" | "column_width" | "fillwidth" | "full" | "mini" | "overflow_fillwidth"
 
 export type ArticleData = {
