+++ src/Components/Publishing/Video/Player/VideoControls.tsx
@@ -1,7 +1,7 @@
 import React, { Component } from "react"
 import styled from "styled-components"
-import { Fonts } from "../Fonts"
-import { IconVideoFullscreen } from "../Icon/IconVideoFullscreen"
+import { Fonts } from "../../Fonts"
+import { IconVideoFullscreen } from "../../Icon/IconVideoFullscreen"
 import { MuteUnmute, MuteUnmuteContainer } from "./MuteUnmute"
 import { PlayPause } from "./PlayPause"
 import { Scrubber } from "./Scrubber"
