+++ src/Components/Publishing/Video/Player/__test__/MuteUnmute.test.tsx
@@ -2,8 +2,8 @@ import { mount } from "enzyme"
 import "jest-styled-components"
 import React from "react"
 import renderer from "react-test-renderer"
-import { IconVideoMute } from "../../Icon/IconVideoMute"
-import { IconVideoUnmute } from "../../Icon/IconVideoUnmute"
+import { IconVideoMute } from "../../../Icon/IconVideoMute"
+import { IconVideoUnmute } from "../../../Icon/IconVideoUnmute"
 import { MuteUnmute } from "../MuteUnmute"
 
 describe("Mute/Unmute Button", () => {
