+++ src/Components/Publishing/Video/Player/__test__/PlayPause.test.tsx
@@ -2,8 +2,8 @@ import { mount } from "enzyme"
 import "jest-styled-components"
 import React from "react"
 import renderer from "react-test-renderer"
-import { IconVideoPause } from "../../Icon/IconVideoPause"
-import { IconVideoPlay } from "../../Icon/IconVideoPlay"
+import { IconVideoPause } from "../../../Icon/IconVideoPause"
+import { IconVideoPlay } from "../../../Icon/IconVideoPlay"
 import { PlayPause } from "../PlayPause"
 
 describe("Play/Pause Button", () => {
