+++ src/Components/Publishing/__stories__/Video.story.tsx
@@ -1,18 +1,21 @@
 import { storiesOf } from "@storybook/react"
 import React from "react"
 import {
-  SeriesArticle,
-  SeriesArticleSponsored,
-  StandardArticle,
   VideoArticle,
-  VideoArticleSponsored
 } from "../Fixtures/Articles"
 import { Media } from "../Fixtures/Components"
+import { VideoPlayer } from "../Video/Player/VideoPlayer"
 import { VideoAbout } from "../Video/VideoAbout"
 import { VideoCover } from "../Video/VideoCover"
-import { VideoLayout } from "../Video/VideoLayout"
 
 storiesOf("Publishing/Video", module)
+  .add("Video Player", () => {
+    return (
+      <div style={{ width: "100vw", height: "100vh" }}>
+        <VideoPlayer {...Media[0]} />
+      </div>
+    )
+  })
   .add("Video Cover", () => {
     return (
       <VideoCover
@@ -28,34 +31,3 @@ storiesOf("Publishing/Video", module)
       />
     )
   })
-  .add("Video Article", () => {
-    return (
-      <VideoLayout article={VideoArticle} />
-    )
-  })
-  .add("Video Article - Series", () => {
-    return (
-      <VideoLayout
-        article={VideoArticle}
-        seriesArticle={SeriesArticle}
-        relatedArticles={[StandardArticle,VideoArticle]}
-      />
-    )
-  })
-  .add("Video Article - Sponsored", () => {
-    return (
-      <VideoLayout
-      article={VideoArticleSponsored}
-      />
-    )
-  })
-  .add("Video Article - Series + Sponsored", () => {
-    return (
-      <VideoLayout
-        article={VideoArticleSponsored}
-        seriesArticle={SeriesArticleSponsored}
-        relatedArticles={[StandardArticle, VideoArticle]}
-      />
-    )
-  })
-
